[hotfix] Create BroadcastITCase.
